+++ CHANGELOG.md
@@ -10,6 +10,8 @@
   - API exposes the CLI parser (`optparse-applicative`) for user-level composition
   - Add `Rib.Parser.Pandoc.getToC` returning rendered Table of contents for a Pandoc document
   - `Rib.Extra.CSS`: add `googleFonts` and `stylesheet`
+- Bug fixes:
+  - `routeUrl`: Fix incorrect substitution of "foo-index.html" with "foo-"
 
 ## 0.7.0.0