+++ CHANGELOG.md
@@ -16,6 +16,7 @@
   - `routeUrl`: Fix incorrect substitution of "foo-index.html" with "foo-"
   - Lucid rendering functions (like `MMark.render`) are now polymorphic in their monad.
   - #122: Fix Pandoc parser never returning metadata
+  - `run` and `runWith` now accept absolute paths (`Path b Dir`).
 
 ## 0.7.0.0