+++ CHANGELOG.md
@@ -17,6 +17,7 @@
   - Lucid rendering functions (like `MMark.render`) are now polymorphic in their monad.
   - #122: Fix Pandoc parser never returning metadata
   - #127: Rib's HTTP server now binds to `127.0.0.1`.
+  - Allow directory listings in HTTP server
   - default.nix: Takes `overrides` and `additional-packages` as extra arguments
   - #130: Prevent unnecessary re-running of Shake action by debouncing fsnotify events