+++ CHANGELOG.md
@@ -6,6 +6,7 @@
   - GHC 8.8
   - pandoc-include-code: 0.5.0.0
   - pandoc-types: 1.20
+  - dhall: 1.13.1
 - New features:
   - API exposes the CLI parser (`optparse-applicative`) for user-level composition
   - Add `Rib.Extra.OpenGraph` for Open Graph protocol